Refactor cache handling: remove obsolete clear-cache functions and update user data overwrite logic
