+++ packages/frontend/containers/settings/OverwriteUserDataContainer.tsx
@@ -1,6 +1,6 @@
 import { FC } from 'react';
 
-import { handleOverwriteUserData } from '$actions/handleOverwriteUserData';
+import { handleOverwriteUserData } from '$actions/user/handleOverwriteUserData';
 import { settingsPaths } from '$constants/settings-paths';
 import { Layout } from '$layouts/Layout';
 import { OverwriteUserData } from '$views/OverwriteUserData';
