+++ packages/frontend/src/actions/user/revalidateUserCache.ts
@@ -1,0 +1,7 @@
+'use server';
+
+import { UserService } from '$ssr/api/UserService';
+
+export const revalidateUserCache = async () => {
+  UserService.revalidateCache();
+};
