--- packages/frontend/src/components/blocks/HardRefreshButton.tsx
@@ -1,12 +1,0 @@
-'use client';
-
-import { Button } from '$elements/Button/Button';
-import { clearAllCaches } from '$ssr/api/clear-cache';
-
-export const HardRefreshButton = () => {
-  return (
-    <Button onClick={() => clearAllCaches()} accentColor="secondary">
-      Clear app cache
-    </Button>
-  );
-};
